backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
psutil>=5.9.0
tqdm>=4.65.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0,<2.0.0
